+++ scripts/scil_NODDI_maps.py
@@ -20,13 +20,14 @@ import amico
 from dipy.io.gradients import read_bvals_bvecs
 import numpy as np
 
+from scilpy.io.gradients import fsl2mrtrix
 from scilpy.io.utils import (add_overwrite_arg,
                              add_processes_arg,
                              add_verbose_arg,
                              assert_inputs_exist,
                              assert_output_dirs_exist_and_empty,
                              redirect_stdout_c)
-from scilpy.gradients.bvec_bval_tools import fsl2mrtrix, identify_shells
+from scilpy.gradients.bvec_bval_tools import identify_shells
 
 EPILOG = """
 Reference:
