+++ scripts/scil_tractogram_commit.py
@@ -84,6 +84,7 @@ import h5py
 import numpy as np
 import nibabel as nib
 
+from scilpy.io.gradients import fsl2mrtrix
 from scilpy.io.streamlines import (reconstruct_streamlines,
                                    reconstruct_streamlines_from_hdf5)
 from scilpy.io.utils import (add_overwrite_arg,
@@ -92,7 +93,7 @@ from scilpy.io.utils import (add_overwrite_arg,
                              assert_inputs_exist,
                              assert_output_dirs_exist_and_empty,
                              redirect_stdout_c)
-from scilpy.gradients.bvec_bval_tools import fsl2mrtrix, identify_shells
+from scilpy.gradients.bvec_bval_tools import identify_shells
 
 
 EPILOG = """
